+++ src/Workspaces.Core/Spelling/SpellcheckAnalyzer.cs
@@ -192,6 +192,18 @@ internal class SpellcheckAnalyzer
                     continue;
                 }
 
+                if (diagnostic.Location.Kind == LocationKind.None)
+                {
+                    if (diagnostic.Properties.TryGetValue("FilePath", out string? filePath))
+                    {
+                        Diagnostic diagnostic2 = Diagnostic.Create(diagnostic.Descriptor, Location.Create(filePath!, default, default), diagnostic.AdditionalLocations, diagnostic.Properties, messageArgs: diagnostic.Properties["Value"]);
+
+                        LogHelpers.WriteDiagnostic(diagnostic2, baseDirectoryPath: Path.GetDirectoryName(project.FilePath), formatProvider: FormatProvider, indentation: "  ", omitSpan: true, verbosity: Verbosity.Normal);
+                    }
+
+                    continue;
+                }
+
                 SpellingDiagnostic spellingDiagnostic = service.CreateSpellingDiagnostic(diagnostic);
 
                 spellingDiagnostics.Add(spellingDiagnostic);